requirements.txt
aioboto3==13.1.0
pandas==2.2.2
s3fs==2024.6.1
rdkit==2023.9.5
tqdm==4.66.4
pyarrow==16.1.0
tenacity==8.4.2
pendulum==3.0.0
botocore==1.34.131
psycopg2==2.9.9
pytest==8.2.2
aiohttp==3.9.5
pydantic==2.7.4
asyncpg==0.29.0
SQLAlchemy==1.4.52
apache-airflow-providers-postgres==5.10.2
apache-airflow-providers-amazon==8.0.0
apache-airflow-providers-slack==8.7.1